got air running

+++ models/variational_autoencoder/VariationalAutoencoder.py
@@ -7,7 +7,7 @@ def log_normal_pdf(x, mu, diag_sigmas):
     return (-D / 2)*tf.log(2*np.pi) - 0.5*tf.reduce_sum(2*tf.log(diag_sigmas), reduction_indices=1) + exp_part
 
 def log_bernoulli_pmf(x, p1):
-    return tf.reduce_sum(tf.log(p1 * x + (1 - p1) * (1 - x)), reduction_indices=1)
+    return tf.reduce_sum(tf.log(p1 * tf.to_float(x) + (1 - p1) * tf.to_float(1 - x)), reduction_indices=1)
 
 def VAE(input, hiddenSize, codeSize, rvType, prefix=''):
     p = prefix
@@ -37,7 +37,7 @@ def VAE_realize(params, random_placeholder, rvType):
     if rvType == 'gaussian':
         return params[0] + tf.mul(params[1], random_placeholder)
     elif rvType == 'bernoulli':
-        return tf.int32(random_placeholder < params[0])
+        return tf.to_int32(random_placeholder < params[0])
     else:
         raise Exception('Unrecognized rvType')
